backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
python-multipart
pdfplumber
python-dotenv
email-validator
supabase
